requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
xlsxwriter
pandas
PyJWT
passlib
python-dotenv